Just before fixing panic: interface conversion
