+++ enterprise/cmd/frontend/internal/batches/webhooks/webhooks_integration_test.go
@@ -20,9 +20,8 @@ func TestWebhooksIntegration(t *testing.T) {
 
 	user := ct.CreateTestUser(t, db, false)
 
-	t.Run("GitHubPushHooks", testGitHubPushHooks(db, user.ID))
-	// t.Run("GitHubWebhook", testGitHubWebhook(db, user.ID))
-	// t.Run("BitbucketServerWebhook", testBitbucketServerWebhook(db, user.ID))
-	// t.Run("GitLabWebhook", testGitLabWebhook(sqlDB))
-	// t.Run("BitbucketCloudWebhook", testBitbucketCloudWebhook(sqlDB))
+	t.Run("GitHubWebhook", testGitHubWebhook(db, user.ID))
+	t.Run("BitbucketServerWebhook", testBitbucketServerWebhook(db, user.ID))
+	t.Run("GitLabWebhook", testGitLabWebhook(sqlDB))
+	t.Run("BitbucketCloudWebhook", testBitbucketCloudWebhook(sqlDB))
 }
