+++ internal/repos/webhooks/github.go
@@ -14,7 +14,7 @@ var (
 	githubEvents = []string{
 		"push",
 	}
-	Url string
+	Url = ""
 )
 
 type SyncWebhook struct {
@@ -31,16 +31,22 @@ func (h *SyncWebhook) handleSyncWebhook(ctx context.Context, extSvc *types.Exter
 	fmt.Println("handleSyncWebhook...")
 	// repo := payload.(*github.PushEvent).GetRepo()
 	// fmt.Printf("repo:%+v\n", repo)
-	var repo api.RepoName
-	repo = "github.com/sourcegraph/sourcegraph"
+	// repoName := repo.Name
+	var repoName api.RepoName
+	repoName = "github.com/sourcegraph/sourcegraph"
 
-	cli := repoupdater.NewClient(Url)
+	var cli *repoupdater.Client
+	if Url == "" {
+		cli = repoupdater.DefaultClient
+	} else {
+		cli = repoupdater.NewClient(Url)
+	}
 
-	res, err := cli.EnqueueRepoUpdate(ctx, repo)
+	res, err := cli.EnqueueRepoUpdate(ctx, repoName)
 	if err != nil {
 		fmt.Println("error in handleSyncWebhook", err)
 	}
-	fmt.Printf("res:%+v\n", res)
+	fmt.Printf("Enqueued:%+v\n", res)
 
 	return nil
 }
