+++ internal/repos/webhooks/testutils.go
@@ -1,0 +1,32 @@
+package syncwebhooks
+
+import (
+	"crypto/hmac"
+	"crypto/sha256"
+	"encoding/hex"
+	"encoding/json"
+	"testing"
+)
+
+func MarshalJSON(t testing.TB, v any) string {
+	t.Helper()
+
+	bs, err := json.Marshal(v)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	return string(bs)
+}
+
+func sign(t *testing.T, message, secret []byte) string {
+	t.Helper()
+
+	mac := hmac.New(sha256.New, secret)
+	_, err := mac.Write(message)
+	if err != nil {
+		t.Fatalf("writing hmac message failed: %s", err)
+	}
+
+	return "sha256=" + hex.EncodeToString(mac.Sum(nil))
+}
